refactor: split resolver.rs into module directory with struct-based design
Extract the monolithic 1133-line resolver.rs into focused modules:
- types.rs: ResolveOptions, ResolverValue, placeholder types, ResObj
- utils.rs: parse_subst_path, lookup_path, deep_merge helpers
- structure_builder.rs: StructureBuilder struct (Pass 1 AST->ResObj)
- substitution_resolver.rs: SubstitutionResolver struct (Pass 2 resolution)
- include_loader.rs: include file loading functions
- mod.rs: public entry point + re-exports + unit tests

Free functions converted to struct methods to reduce parameter passing.
No logic changes — pure structural refactoring. All 220 tests pass.
